Add files via upload
The intelligence center has three types of workers: one operations manager and a fixed number of telegraphers and cryptographers.

Each telegraph operator listens to the channel on which encrypted messages arrive. When he receives one message, he gives it to the monitor. He does this until no more messages arrive (e.g.when he gets a message signaling END). 

The messages in the monitor are encrypted and must be decrypted by the cryptographers.
Each cryptographer retrieves an encrypted message from the monitor, decrypts it, and sends the decrypted message to a new monitor. He does this until there are no more encrypted messages, and no more can come (e.g., when all telegraph operators have finished, they give one special message to the monitor that says NO MORE MESSAGES).

The operations manager retrieves decrypted messages from the second monitor. He brings together the messages and sorts them based on channel and sequence number. When he knows that no new messages can arrive on this monitor (the cryptographers report in the same way as the telegraph operators above) he prints out the messages for each channel in sorted order.

Each person — the telegraph operators, the cryptographers, and the operations manager — is a thread. All the threads must run in parallel. This means that several channels can be listened to simultaneously, several messages can be decrypted simultaneously, and the program runs faster. Both monitors (one for encrypted
and one for decrypted messages) must be protected with a lock to avoid multiple threads
trying to modify them simultaneously.